GPU: VertexFormat: Use new data types
This prevents the use of unaligned data types in
vertex formats.

This simplify the `GPUVertexFormat` class a lot as
we do not need packing shenanigans anymore and just
compute the vertex stride.

The old enums are kept for progressive porting of the
backends and user code.

This will break compatibility with python addons.

TODO:
- [ ] Deprecation warning for PyGPU (4.5)
- [ ] Error handling for PyGPU (5.0)
- [ ] Backends
  - [x] Metal
  - [ ] OpenGL
  - [ ] Vulkan