Refactor: remove ArmatureBoneDrawStrategy
Soon the draw type will be per-bone, so you can't just use the strategy
object for the whole armature. Besides, the "strategy pattern" does
not seem to do all that much here; it saves one switch statement
and several ifs, at the cost of 150 lines of code verbosity.

Should be no functional changes