Fix #138388: Use grid voxel corners as value locations like OpenVDB.
Blender grid rendering interprets voxel transforms in such a way that the voxel
values are located at the center of a voxel. This is inconsistent with OpenVDB
where the values are located at the lower corners for the purpose or sampling
and related algorithms.

While it is possible to offset grids when communicating with the OpenVDB
library, this is also error-prone and does not add any major advantage.
Every time a grid is passed to OpenVDB we currently have to take care to
transform by half a voxel to ensure correct sampling weights are used that match
the density displayed by the viewport rendering.

This patch changes volume grid generation, conversion, and rendering code so
that grid transforms match the corner-located values in OpenVDB.

- The volume primitive cube node aligns the grid transform with the location of
  the first value, which is now also the same as min/max bounds input of the
  node.
- Mesh<->Grid conversion does no longer require offsetting grid transform and
  mesh vertices respectively by 0.5 voxels.
- Texture space for viewport rendering is offset by half a voxel, so that it
  covers the same area as before and voxel centers remain at the same texture
  space locations.

- [x] OpenVDB grid file import does not require changes. This was generally
  incorrect already due to file grids using corner locations, and now displays
  files correctly.
- [x] Cycles does not render generated grids from object data and does its own
  conversion for file imports, so is not affected by this patch.
- [ ] The grid bounds have now changed and don't include the 0.5 voxel padding
  on the left side any more. This could be accounted for when converting grid
  bounds to object bounds. However, grid bounds are not well defined already
  (not accurate on the right side, sampling will yield non-zero values outside
  current bounds). This needs more discussion.