Nodes: Add Subtype, Min, and Max properties to RGBA sockets
This patch extends the RGBA node sockets to allow setting subtypes and
minimum and maximum values much like Vector sockets for instance. This
is needed by the compositor since some color sockets typically need a
[0, 2] range and some other sockets are gamma corrected, and this need a
PROP_COLOR_GAMMA subtype.

The main complication I see in this patch is the fact that color sockets
can't really have PROP_NONE subtype, they can either be PROP_COLOR or
PROP_COLOR_GAMMA. But PROP_NONE is taken as a special value in many
places of code to indicate a base socket type. For that reason, it is
necessarily to have a PROP_NONE variant of the color socket, so we
register the PROP_COLOR variant as a PROP_NONE, but otherwise use the
PROP_COLOR subtype internally.