Physics: Decrease thickness of Smoke surface Emission to better fit small objects
The Smoke Surface Emission is height of the smoke from mesh surface.
It is measured in Domain Grid Units, and the Grid unit is big (compared to mesh),
the smoke emission is huge.

To fix that the Surface Emission was decreased from 1.5 to 1.0.
As a result it is possible to emit smaller smokes (width 2, with previous implementation 3).

Further decreasing default value of Surface Emission is not recommended,
as in some cases (too small mesh compared to Grid Unit) will result not emitted smoke.

Fixes: #132613