Fix #111152: GHOST: Catch WinTab exception on `WTOpen`
The WinTab driver from some vendors (like Huion) sometimes can get
into a corrupted state in which `WTOpen` call will throw an
exception internally. This needs to be caught in order for blender
to resume normal operation. This PR also shows a message box when
this happens, which tells the user to restart tablet driver.