Tests: Add initial render procedural tests
This commit adds a new render test folder for the render
procedural feature of the mesh sequence cache modifier, currently used
by Cycles.

In that folder a single test has been added, importing an Alembic file
that contains a simple 3 frame animation of a moving, rotating,
scaling, and deforming mesh.

The test is setup to test:
- The importing of mesh objects using render procedural
- And to test to make sure frame offsets work correctly.
  - There was a bug report it wasn't working properly in (#134043)