Fix #138236: Vulkan: destroyed pipeline layouts could be reused
Due to an oversight in the pipeline pool/shader, pipelines that were
invalid could be reused. This happened more often when using the
compositor as it recreates shaders over and over. It also depended on
driver handle reuses on how often artifacts could have been seen.

This is fixed by discarded any associated pipeline when the pipeline
layout is discarded.