Geometry Nodes: "Rename Attributes" node for bulk-renaming attributes
This node allows bulk-renaming named attributes.
This is in-line with the "Remove Attribute" node.
It uses roughly the same wildcard syntax, adjusted for find/replace.
This is useful to replicate i.e. the effects of the Mirror modifier.
It is also useful to isolate specific groups from deep node graphs.