BLF: Synthesize Text Object Space Character if Not Available
For Text Objects that use a custom font we no longer take characters
from the UI Fallback stack when characters are not found. In this case
though the font itself does not contain a space character. Rather than
requiring breaking up the work into separate objects this PR creates a
space character in a predicable way. It loads a space from the stack
but then alters the width  - a full em for monospaced fonts, 25% of em
width for proportional. Why do this for space and not other characters?
Many users will not understand that a space character is part of the
font design and could consider it part of formatting and spacing, like
carriage return or tab.