Fix: Nodes: Missing builtin categories for handling asset catalogs
The current behavior for the Node Editor's Add
menu is to list the user's asset catalogs at the
root level of the menu. However, an exception is
made if the catalog has the same name as a builtin
menu. In such case, the asset catalog is
appended to that builtin menu, instead of being
drawn at the root level.

How this is currently implemented right now is to
have a hard-coded set of builtin menus in
`add_menu_assets.cc`. Certain menus are missing
from this set, particularly newer ones, which
may be an oversight during the creation of these
menus.

This patch adds those menus to that list, which are as follows:
- Input/Gizmo
- Input/Import
- Grease Pencil (and all its submenus)
- Texture (in Compositor Nodes)