GPU: Remove 32 bit int to float fetch mode combination
This commit finishes removing the uses of the 32 bit integer to float
vertex buffer fetch mode. Previous commits noted below already started
that process. The last usage was geometry attributes. Now integers are
converted to floats as part of the existing upload process.

The change makes the Vulkan vertex buffer type conversion unused, so
it's removed. That's nice because Vulkan vertex buffers go from 1040 to
568 bytes in size and have significantly less overhead on creation.

Related to:
- 153abc372e914b92bb960cf17482fd1956a5ec96
- 1e1ac2bb9b3bbef0443038af7f3a9e4446fe6201
- 617858e453c3496a1a403b5263991c3326e42131