Fix: Geometry Nodes: Asset catalog named "Unassigned" does not show up in menu
---

The Node Editor Add menu has filters for not
drawing asset catalogs in the root level.
This is done so that catalogs that share names
with built-in menus are appended to said menus.

There is no "Unassigned" built-in menu, just a
root level asset menu for unassigned assets.
This menu does not currently have the behavior
to support appending catalogs to it, so assigning
assets to a catalog named "Unassigned" renders
said assets unaccessible in the Add Menu.

This patch removes that filter as it was deemed
better to just allow the user to have a root
level catalog menu named "Unassigned" instead
of mushing it with the menu for unassigned assets.