Cycles: Use CUDA 11 to compile PTX kernels
This commit makes it so CUDA 11 is used to compile PTX CUDA kernels
(E.g. compute_75). This is done to ensure the GPU driver requirements
for users of the PTX kernels (primarily server GPU users like A100)
low.

This is important because in some situations, it's either impossible
(E.g. Renting certain cloud services), or difficult to update the GPU
drivers on some machines and we want to make sure the PTX kernels
are usable by as many people as possible

Original Author: Sergey Sharybin