UV: face-only selection now uses its own "uvselect" data
Unfortunately there are subtle differences between sync-select
in face selection mode with/without separate "uvselect" data.

- It's not possible to "split" the selection.
- When switching from face to edge/vertex modes,
  the absence of UV selection data means the UV editor may
  select adjacent UV's of islands which were not selected
  as part of the face.

Even though these can be overlooked, I'm not so keen on making
face-only selection as a special-case which needs to be accounted for
separately from other selection modes.

It also seems likely there will be other, similar situations that
arise where there are minor differences which users are likely
to notice as bugs/glitches.