Fix: Grease Pencil: Tree Node cache order
The ordering of tree nodes is meant to be bottom-up, but the
`nodes()` cache ordering was not quite right.

E.g. in this example:
```
Group
   Layer 1
   Layer 2
Layer 3
```
the resulting `nodes()` output would be:
```
Layer 3
Group
Layer 2
Layer 1
```
So the `Group` node was incorrectly ordered before the children.

Now, this results in the (expected) result:
```
Layer 3
Layer 2
Layer 1
Group
```