correct previous commit
We actually want the `CfraElem` `cfra` to be in "global" scene time (not
"raw" FCurve data) because multiple data-blocks (mix of NLA tweak and
untweaked) might be shown in the Dopesheet/Graph Editors. If we store
"raw/local" frame data in `CfraElem` there is no way to NLA-map that to
other data-blocks "strip-time/local" keyframes.

So in order to do that, we perform the NLATIME_CONVERT_MAP in
`bezt_to_cfraelem`