Fix #138793: Grease Pencil: Tag ID update after conv from legacy curves
Converting from legacy curves to grease pencil with "Keep Original"
option enabled can cause the new grease pencil ID to not be evalueated,
leading to an situation where a evaluated grease pencil object would
point to the original legacy curve ID as its object data. This fix added
a depsgraph tag so it ensures the COW ID is the latest.