Fix #138933: Support assigning a default value to menus when wrapping node groups.
The modifier input value for menu items is an int property, which is not compatible
with the enum property used in node group socket definitions for menus.
Wrapping a node group with a menu will try to assign the int value from the modifier
directly to the enum, which expects a string identifier.

To aquire the matching string identifier, use the UI properties in the modifier RNA,
which contain the complete enum item definition.