Fix: `RNA_property_string_get()` always returns zero-length string
The variant of `RNA_property_string_get()` that returns a `std::string`
always returned a string with zero length. The issue is that it was
only calling `reserve()`, which ensures an underlying buffer with enough
*capacity*, but does not set the string length. The property's string
value would then be correctly copied to the reserved buffer, but the
string length would remain zero.

This commit fixes the issue by replacing the call to `reserve()` with
a call to `resize()`, which both ensures the buffer size *and* sets the
string length.