ANIM: optimize proportional editing many keys in the Graph Editor
Part one.

We have to determine distances for proportional editing for all three
`TransData` (representing the key as well as both handles). The way
#bezt_to_transdata sets up the `TransData`, the `td->center[0]` will
always be based on the **key** (bezt->vec[1]) which means that
#graph_key_shortest_dist will return the same for all of them and we can
reuse that (expensive) result if needed.

Depending on the situation, we can save two thirds of the time spent
here... (42sec >> 14sec before we can actually start transforming keys
in the file from #139042)

Might be worth looking into using a 2D KDTree in the future as well.

(also corrects a code comment about what is actually done in a
particular code path)

Noticed while looking into #139042