WIP: Assets: add OpenAPI data model generator
Add a Python code generator that takes an OpenAPI definition and outputs
the corresponding data model as Pydantic classes.

The generator can be run via `make generate_datamodels`. This will
create a Python virtualenv, install the dependencies it needs, and then
run the generator within that virtualenv.

The virtualenv is created in a temporary directory, and is not cleaned
up after running. This means that subsequent runs will just use it
directly, instead of reinstalling dependencies on every run.

**Open Questions:**

- VirtualEnv: for this one, or more global?

  An open question is whether this is the right approach for Blender, or
  whether we want to have a more general virtualenv management for more
  Python development dependencies (like Sphynx for the docs generator).

- OpenAPI spec: in source or separate?

  Currently the OpenAPI spec file is stored in
  `scripts/modules/bl_assets/blender_asset_library_openapi.yaml`. I
  think for now this is fine, but maybe at some point, when this is also
  used by the server, it might become more interesting to place the file
  somewhere less Blender (the program) specific.

# Conflicts:
#	.gitattributes