Move options into DownloaderOptions class
This makes it easier to pass options from the BackgroundDownloader
initializer to the ConditionalDownloader it wraps.

Because this adds yet another class, I've refactored the imports to only
import the module, and not all the used names from that module.'

Also this moves the creation of the subprocess object to the `.start()`
function (as mentioned earlier in a TODO comment).