Fix #139042: Graph Editor proportional editing can corrupt key values
... if integer-only FCurves are present

Values could "randomly" be set to the closest integer (most noted on
rotation values since these are in radians).

The underlying issue is in fact quite worrying (and it would probably be
good if this doesnt cause issues elsewhere as well).

So, the underlying problem is that `TransData` gets sorted by
`sort_trans_data_dist_container` for proportional editing whereas
`TransData2D` isnt... And... we usually iterate over both of them with
the assumption these are ordered to be corresponding, but in fact they
are not.

In the particular case of the Graph Editor, this gets to be a problem
since we check a `TransData` flag (`TD_INTVALUES`) in
`flushTransGraphData`, then act on the ~corresponding~ wrong
`TransData2D` -- in this case: make it the closest INT -- which is super
bad. Cancelling the transform does not bring it back, undo doesnt
either.

The ~fix~ workaround for the particular Graph Editor case would be to
**not** flag the `TransData`, but the `TransData2D` directly (so we can
check that in `flushTransGraphData`), this way the fact that `TransData`
and `TransData2D` are not in sync there is not important anymore.

Still, like I said, the "cleaner" fix would probably be to make sure
that when we sort, we sort **everything** we later iterate over the
same. (This can be done in a different commit though, moving the
`TransData` flag to a `TransData2D` property does not hurt either way)