Tests: Add UI test that cycles through all default workspaces
While individual modes have UI tests related to undo, this test and this
new file is intended to be a set of very broad tests that catch the most
egreious errors that cause crashing on start up.

Running this test takes 2~3 seconds currently as it cycles through the
11 default workspaces and checks the name of each to ensure that it has
loaded correctly.