Fix: Graph Editor box select selects non-visible handles
We should never select something that is hidden in the viewport.

Generally, there can be multiple reasons why keyframe handles are not
shown in the Graph Editor:
- the `View` > `Show Handles` / `Only Show Selected Keyframes Handles`
options is disabled
- a keyframe `Interpolation`  method is set to `Linear`

This PR corrects the first (so it now combines the existing operator
include_handles option with the mentioned options).

NOTE: same check is already done for Circle and Lasso select.

Part of #139314