Fix: Graph Editor box select menu entries
Currently we have two entries in the Select menu:
- Box Select
- Box Select (Include Handles)

But since b037ba2665f4, Include Handles is the default, so both entries
behave exactly the same.

To resolve, make it:
- Box Select
- Box Select (Skip Handles)

with the second one setting the "include_handles" option to False.

NOTE: behavior of "include_handles" option False might be a bit flaky in
certain situations as well (can be checked later), better to actually
have two entries that represent what the code does.

Noticed while checking on #139314