VSE: Add options to show origins and all origins in overlays
When "Origins" is enabled, the origin points of selected strips are
shown.

When "Origins (All)" is also enabled, origins of all strips are
shown in the preview, not just the selected ones."Origins (All)" only
works if "Origins" is also enabled.

When "Origins" is disabled origins are shown only when needed like
when moving stuff.